Refactored dsp into dsp module.
